+++ src/core/distributions/ZTPoisson.h
@@ -21,19 +21,19 @@
 
 namespace transmission_nets::core::distributions {
     template<int MAX_COUNT>
-    class ZTPoisson : public computation::Computation<std::array<long double, MAX_COUNT + 1>>,
+    class ZTPoisson : public computation::Computation<std::array<double, MAX_COUNT + 1>>,
                       public abstract::Observable<ZTPoisson<MAX_COUNT>>,
                       public abstract::Cacheable<ZTPoisson<MAX_COUNT>>,
-                      public abstract::Checkpointable<ZTPoisson<MAX_COUNT>, std::array<long double, MAX_COUNT + 1>> {
+                      public abstract::Checkpointable<ZTPoisson<MAX_COUNT>, std::array<double, MAX_COUNT + 1>> {
         using p_ParameterDouble = std::shared_ptr<core::parameters::Parameter<double>>;
 
     public:
         explicit ZTPoisson(p_ParameterDouble mean) noexcept;
 
-        std::array<long double, MAX_COUNT + 1> value() noexcept;
+        std::array<double, MAX_COUNT + 1> value() noexcept;
 
     private:
-        friend class abstract::Checkpointable<ZTPoisson<MAX_COUNT>, std::array<long double, MAX_COUNT + 1>>;
+        friend class abstract::Checkpointable<ZTPoisson<MAX_COUNT>, std::array<double, MAX_COUNT + 1>>;
         friend class abstract::Cacheable<ZTPoisson<MAX_COUNT>>;
 
         p_ParameterDouble mean_;
@@ -41,7 +41,7 @@ namespace transmission_nets::core::distributions {
 
     template<int MAX_COUNT>
     ZTPoisson<MAX_COUNT>::ZTPoisson(p_ParameterDouble mean) noexcept : mean_(std::move(mean)) {
-        this->value_[0] = -std::numeric_limits<long double>::infinity();
+        this->value_[0] = -std::numeric_limits<double>::infinity();
         mean_->registerCacheableCheckpointTarget(this);
         mean_->add_post_change_listener([=, this]() {
             this->setDirty();
@@ -51,10 +51,10 @@ namespace transmission_nets::core::distributions {
     }
 
     template<int MAX_COUNT>
-    std::array<long double, MAX_COUNT + 1> ZTPoisson<MAX_COUNT>::value() noexcept {
+    std::array<double, MAX_COUNT + 1> ZTPoisson<MAX_COUNT>::value() noexcept {
         if (this->isDirty()) {
             double lambda = mean_->value();
-            long double denominator = 0.0;
+            double denominator = 0.0;
             for (int j = 1; j < MAX_COUNT + 1; ++j) {
                 this->value_[j] = j * std::log(lambda) - std::log(std::exp(lambda) - 1) - std::log(boost::math::factorial<double>(j));
                 denominator += exp(this->value_[j]);
